chore(upgrades): converts upgrade related javascript into an AMD module
